==> gold_classification/3_gold_llm_summarizer.ipynb <==
from pyspark.sql import functions as F

df_gold_excluded = spark.table("gold.conversations_excluded")
# display(df_gold_excluded)

df_gold_excluded = df_gold_excluded.withColumn(
    "convo_summary",
    F.expr("ai_summarize(combined_text, 35)")
)

df_gold_excluded.write.mode("overwrite").saveAsTable("gold.conversations_summarized")
# display(df_gold_excluded)
