==> gold_classification/5_gold_semantic_classification.ipynb <==
%pip install --upgrade --force-reinstall databricks-vectorsearch
dbutils.library.restartPython()

## Load Vector Index

from databricks.vector_search.client import VectorSearchClient

vsc = VectorSearchClient(disable_notice=True)

vector_search_endpoint_name = "widlchat"
vsc.get_endpoint(
  name=vector_search_endpoint_name
)

# Overview tab for the vector index.
vs_index_fullname = "workspace.gold.idx_convos_classified"
index = vsc.get_index(endpoint_name=vector_search_endpoint_name,index_name=vs_index_fullname)
index.describe()

## Semantic classification function

from pyspark.sql import functions as F
from pyspark.sql import types as T

# Load data and classify
df_gold_summarized = spark.table("gold.conversations_summarized")

df_gold_summarized_sampled = (
    df_gold_summarized
    .orderBy(F.rand())
    .limit(5000)
)

df_gold_summarized_sampled.createOrReplaceTempView("sampled_src")

df_semantic_classified = spark.sql("""
SELECT
  src.conversation_id,
  src.ISID,
  src.country,
  src.combined_text,
  src.char_count,
  src.convo_summary,

  -- semantic hit → category, else empty
  CASE
    WHEN vs.search_score >= 0.75 THEN vs.top_category
    ELSE ''
  END AS top_category,

  -- semantic hit → similarity score, else 0.0
  CASE
    WHEN vs.search_score >= 0.75 THEN vs.search_score
    ELSE 0.0
  END AS similarity_score,

  -- semantic hit → 'semantic', else empty
  CASE
    WHEN vs.search_score >= 0.75 THEN 'semantic'
    ELSE ''
  END AS classification_method

FROM sampled_src src

LEFT JOIN LATERAL (
  SELECT
    result.top_category,
    result.search_score
  FROM vector_search(
    index => 'workspace.gold.idx_convos_classified',
    query_text => src.combined_text,
    num_results => 1
  ) AS result
) vs
""")

# spark.sql("DROP TABLE gold.conversations_semantic_classified")
df_semantic_classified.write.mode("overwrite").saveAsTable(
    "gold.conversations_semantic_classified"
)


spark.table("gold.conversations_semantic_classified") \
     .groupBy("classification_method") \
     .count() \
     .orderBy(F.desc("count")) \
     .show()
